refactor: split token provider
